+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Navigation from './App';
+import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(
-  <Navigation />,
+  <App />,
   document.getElementById('root')
 );
 registerServiceWorker();
